Move markdown validation to validators and refactor loadCollections for modularity
