+++ util/validators.ts
@@ -1,3 +1,7 @@
+import greyMatter from 'gray-matter'
+
+import FrontMatter from './FrontMatter'
+
 export function isString(value: unknown): value is string {
     return typeof value === 'string'
 }
@@ -28,4 +32,20 @@ export function normalizeString(rawString: unknown): string {
         return rawString.trim()
     } 
     throw new Error('Front-matter: Value must be a string.')
+}
+
+export async function normalizeFrontMatter(rawFrontMatter: ReturnType<typeof greyMatter>['data']): Promise<FrontMatter> {
+    return {
+        title: normalizeString(rawFrontMatter.title) ?? normalizeString(rawFrontMatter.alternateTitle) ?? 'Untitled Document',
+        alternateTitle: rawFrontMatter.alternateTitle ? normalizeString(rawFrontMatter.alternateTitle) : '',
+        tags: normalizeTags(rawFrontMatter.tags),
+        published: isBoolean(rawFrontMatter.published) ? rawFrontMatter.published : true,
+        unlisted: isBoolean(rawFrontMatter.unlisted) ? rawFrontMatter.unlisted : false,
+        author: rawFrontMatter.author ? normalizeString(rawFrontMatter.author) : '',
+        image: rawFrontMatter.image ? normalizeString(rawFrontMatter.image) : '',
+        imageAlt: rawFrontMatter.imageAlt ? normalizeString(rawFrontMatter.imageAlt) : '',
+        socialImage: rawFrontMatter.socialImage ? normalizeString(rawFrontMatter.socialImage) : '',
+        description: rawFrontMatter.description ? normalizeString(rawFrontMatter.description) : '',
+        type: normalizeString(rawFrontMatter.type)
+    }
 }
